updated my branch with latest working version

+++ app.py
@@ -1,40 +1,67 @@
-from flask import Flask, request, make_response, jsonify, redirect, Response, flash
+####################
+# Course: CSE138
+# Date: Spring 2020
+# Author: Alan Vasilkovsky, Mariah Maninang, Bradley Gallardo, Omar Quinones
+# Assignment: 3
+# Description: 
+###################
+
+from flask import Flask, request, make_response, jsonify, Response
 import os
 import requests
-import json, time
-import sys
+import json
 app = Flask(__name__)
 
-socket_address = os.getenv('SOCKET_ADDRESS')
-#socket_address = "10.10.0.2:8085"
-myIP = socket_address
-
-viewVar = os.getenv('VIEW')
-#viewVar = "10.10.0.2:8085,10.10.0.3:8085,10.10.0.4:8085"
-view = viewVar.split(',')
-
 key_value_store = {}
-
+myIP = os.getenv('SOCKET_ADDRESS')
 vectorClock = {"10.10.0.2:8085": 0, "10.10.0.3:8085":0, "10.10.0.4:8085":0}
 requestQueue = {}
 
-@app.route('/test')
-def test():
-    return request.remote_addr, 200
+viewVar = os.getenv('VIEW')
+view = viewVar.split(',')
+
+def wakeup():
+
+    global key_value_store
+    for ip in [i for i in view if i != myIP]:
+        try:
+            other_kvs = requests.get('http://%s/wake' % ip, timeout=1).json()
+
+            if key_value_store != other_kvs:
+            
+                for key in other_kvs:
+
+                    if key not in key_value_store:
+
+                        key_value_store[key] = other_kvs[key]
+                        key_value_store[key]["causal-metadata"][myIP] += 1
+                        vectorClock[myIP] += 1
+
+                    elif key in key_value_store and key_value_store[key]["causal-metadata"][myIP] < other_kvs[key]["causal-metadata"][ip]:
+
+                        key_value_store[key] = other_kvs[key]
+                        key_value_store[key]["causal-metadata"][myIP] += 1
+                        vectorClock[myIP] += 1
+
+                    # update_other_replica_vc()
+
+            return
+        except (requests.exceptions.Timeout, requests.exceptions.ConnectionError):
+            continue
 
 @app.route('/key-value-store-view', methods=['GET', 'PUT', 'DELETE'])
 def view_operations():
     global view
 
+
     if request.method == 'GET':
-        old_view = view
-        sender = request.remote_addr + ":8085"
         broadcast(request)
 
-        return make_response('{"message":"View retrieved succesfully","view":%s}' % json.dumps(view), 200)
+        return make_response('{"message":"View retrieved successfully","view":%s}' % json.dumps(view), 200)
 
 
     elif request.method == 'PUT':
+        old_view = view
         sender = request.remote_addr + ":8085"
 
         address_to_be_added = request.json['socket-address']
@@ -69,7 +96,7 @@ def view_operations():
 def broadcast(request):
     global view
 
-    broadcast_range = [ip for ip in vectorClock if ip != socket_address] 
+    broadcast_range = [ip for ip in vectorClock if ip != myIP] 
     print("broadcast range: {}".format(broadcast_range))
 
     for ip in broadcast_range:
@@ -106,10 +133,11 @@ def broadcast(request):
                     requests.delete('http://{}/key-value-store-view'.format(replica), json={"socket-address": ip})
 
             #put additional active replicas to other replicas' views
-            put_broadcast_range = [ip for ip in view if ip != socket_address]
+            put_broadcast_range = [ip for ip in view if ip != myIP]
             for replica in put_broadcast_range:
                 for ip in view:
                     res = requests.put('http://{}/key-value-store-view'.format(replica), json={"socket-address": ip})
+                    return res
             
 
 
@@ -121,7 +149,7 @@ def broadcast(request):
                 view.remove(ip)
                 print(view)
                 
-                del_broadcast_range = [ip for ip in view if ip != socket_address]
+                del_broadcast_range = [ip for ip in view if ip != myIP]
                 for replica in del_broadcast_range:
                     try:
                         requests.delete('http://{}/key-value-store-view'.format(replica), json={"socket-address": ip}, timeout=1)
@@ -150,60 +178,45 @@ def broadcast(request):
 @app.route('/key-value-store/<key>', methods=['GET','PUT','DELETE'])
 def kvs(key):
     global view
-    global key_value_store
-
 
     sender = request.remote_addr + ":8085"
 
     if sender not in vectorClock:
         res = requests.get('http://{}/key-value-store-view'.format(myIP), headers = request.headers).json()
         view = res["view"]
-        
+
+        wakeup()
+
 
     if request.method == 'GET':
-
-        if key not in key_value_store or key_value_store[key] is None:
-            return make_response('{"message":"Key does not exist"}', 404)
-
+    
+        if key_value_store[key] is None or key not in key_value_store:
+            return make_response('{"doesExist":false,"error":"Key does not exist","message":"Error in GET"}', 404)
         else:
             causal = json.dumps(key_value_store[key]["causal-metadata"])
             value = key_value_store[key]["value"]
-            return make_response('{"message":"Retrieved successfully", "causal-metadata":%s, "value":"%s"}' % (causal, value), 200)
-  
+
+            return make_response('{"doesExist":true,"message":"Retrieved successfully","causal-metadata":%s,"value":"%s"}' % (causal, value) , 200)
 
 
     if request.method == 'PUT':
+
         causal = request.json["causal-metadata"]
         value = request.json
 
-        if key in key_value_store:
-            successMsg, statusCode = "Updated successfully", 200
-        else:
-            successMsg, statusCode = "Added successfully", 201
+        keyAlreadyExists = True if key in key_value_store else False
 
         if causal == "":
 
             key_value_store[key] = request.json
             vectorClock[myIP] += 1
 
-            #check requests on hold
-            if requestQueue != {}:
-                done = False
-                while done is not True:
-                    done = checkRequestQueue()
-
-            kvs_broadcast(sender, key, value, request.method)
-
-            vectorClockJson = json.dumps(vectorClock)
-            return make_response('{"message":"%s", "causal-metadata": %s}' % (successMsg, vectorClockJson), statusCode)
-
         else:
 
             if causal[myIP] > vectorClock[myIP]:
                 requestQueue[key] = request.json
                 requestQueue[key]["method"] = request.method
-                requestQueue[key]["sender"] = request.remote_addr + ":8085"
-                return ('{"message": "Request placed in a queue", "request": %s}' % json.dumps(requestQueue[key]), 200)
+                return make_response('{"message": "Request placed in a queue", "request": %s}' % json.dumps(requestQueue[key]), 200)
 
             else:
 
@@ -216,20 +229,27 @@ def kvs(key):
                     takeMaxElement(causal)
                     vectorClock[myIP] += 1
 
-                #check requests on hold
-                if requestQueue != {}:
-                    done = False
-                    while done is not True:
-                        done = checkRequestQueue()
+        key_value_store[key]["causal-metadata"] = vectorClock
+        kvs_broadcast(sender, key, value, request.method)
+        update_other_replica_vc(sender)
 
-                kvs_broadcast(sender, key, value, request.method)
+        #check requests on hold
+        if requestQueue != {}:
+            done = False
+            while done is not True:
+                done = checkRequestQueue()
 
-                vectorClockJson = json.dumps(vectorClock)
-                return make_response('{"message":"%s", "causal-metadata": %s}' % (successMsg, vectorClockJson), statusCode)
+        vectorClockJson = json.dumps(vectorClock)
+        if keyAlreadyExists ==  True:
+            return make_response('{"message":"Updated successfully", "causal-metadata": %s}' % vectorClockJson, 200)
+        else:
+            return make_response('{"message":"Added successfully", "causal-metadata": %s}' % vectorClockJson, 201)
 
             
     if request.method == 'DELETE':
+
         causal = request.json["causal-metadata"]
+        sender = request.remote_addr + ":8085"
         value = request.json
 
         if key not in key_value_store:
@@ -238,7 +258,6 @@ def kvs(key):
         elif causal[myIP] > vectorClock[myIP]:
             requestQueue[key] = request.json
             requestQueue[key]["method"] = request.method
-            requestQueue[key]["sender"] = request.remote_addr + ":8085"
             return make_response('{"message": "Request placed in a queue", "request": %s}' % json.dumps(requestQueue[key]), 200)
 
         elif key_value_store[key] is None:
@@ -254,16 +273,19 @@ def kvs(key):
                 takeMaxElement(causal)
                 vectorClock[myIP] += 1
 
+            key_value_store[key]["causal-metadata"] = vectorClock
+            kvs_broadcast(sender, key, value, request.method)
+            update_other_replica_vc(sender)
+
+
             #check requests on hold
             if requestQueue != {}:
                 done = False
                 while done is not True:
                     done = checkRequestQueue()
 
-            kvs_broadcast(sender, key, value, request.method)
-
             vectorClockJson = json.dumps(vectorClock)
-            return make_response('{"doesExist":true,"message":"Deleted successfully", "causal-metadata":"%s"}' % vectorClockJson, 200) 
+            return make_response('{"message":"Deleted successfully", "causal-metadata":%s}' % vectorClockJson, 200) 
 
     
 def checkRequestQueue():
@@ -273,70 +295,70 @@ def checkRequestQueue():
         causal = requestQueue[key]["causal-metadata"]
         if causal[myIP] <= vectorClock[myIP]:
 
-            sender = requestQueue[key].pop("sender")
-            method = requestQueue[key].pop("method")
-
-            if method == 'PUT':
+            if requestQueue[key]["method"] == 'PUT':
                 key_value_store[key] = requestQueue[key]
+                method = 'PUT'
 
-            elif method == 'DELETE':
+            elif requestQueue[key]["method"] == 'DELETE':
                 key_value_store[key] = None
+                method = 'DELETE'
 
+            sender = request.remote_addr + ":8085"
             value = key_value_store[key]
 
             del requestQueue[key]
             vectorClock[myIP] += 1
-
             kvs_broadcast(sender, key, value, method)
 
             return False
         
     return True
     
-
 def takeMaxElement(causal):
     for ip in vectorClock:
         if causal[ip] > vectorClock[ip]:
             vectorClock[ip] = causal[ip]
-
+        
 
 def kvs_broadcast(sender, key, value, method):
-    global view
-
+    
     #if sender is not a replica, broadcast
     if sender not in vectorClock:
-        broadcast_range = [ip for ip in view if ip != socket_address] 
-        for ip in broadcast_range:
+        for ip in view:
+            if ip != myIP:
 
-            address = "http://" + ip + "/key-value-store/" + key
-            value["causal-metadata"] = key_value_store[key]["causal-metadata"]
+                address = "http://" + ip + "/key-value-store/" + key
+                value["causal-metadata"] = vectorClock
 
-            if method == 'PUT':
-                response = requests.put(address, headers=request.headers, json=value)
+                if method == 'PUT':
+                    response = requests.put(address, headers=request.headers, json=value)
 
-            elif method == 'DELETE':
-                response = requests.delete(address, headers=request.headers, json=value)
+                elif method == 'DELETE':
+                    response = requests.delete(address, headers=request.headers, json=value)
 
-            responseJson = response.json()
-            newVector = responseJson["causal-metadata"]
-            takeMaxElement(newVector)
+                responseJson = response.json()
+                newVector = responseJson["causal-metadata"]
+                takeMaxElement(newVector)
 
 
+def update_other_replica_vc(sender):
+    if sender not in vectorClock:
+        ip_to_get_vc = [ip for ip in view if ip != myIP]
+
+        for ip in ip_to_get_vc:
+            requests.put('http://{}/send-vc'.format(ip), headers = request.headers, json = {"vector-clock": vectorClock})
+    return
+
+
+@app.route('/send-vc', methods = ['PUT'])
+def send_vc():
+    other_vc = request.json["vector-clock"]
+    takeMaxElement(other_vc)
+    return
 
 @app.route('/status', methods = ['GET'])
 def status():
-    global view
-
     return make_response('{"status": "alive"}')
-
-# @app.route('/get-vc', methods = ['GET'])
-#     return {"vector-clock": json.dumps(vectorClock)}
-
-# @app.route('/update-kvs', methods = ['PUT'])
-# def update_kvs():
-#     global key_value_store
-
-#     key_value_store = request.json["key-value-store"]
 
 @app.route('/wake', methods=['GET'])
 def wake():
@@ -344,10 +366,3 @@ def wake():
 
 if __name__ == '__main__':
     app.run(debug=True, host='0.0.0.0', port=8085)
-    for ip in [i for i in view if i is not socket_address]:
-        try:
-            key_value_store = requests.get('http://%s/wake' % ip, timeout=1)[0]
-            break
-        except requests.exceptions.Timeout:
-            continue
-
